Automapper issues
My previous change where adding IEquatable<Anchor> to the Model.Anchor exposed a bug in AutoMapper when hosted under .NET 7 Runtime.  This was resolved by moving from AutoMapper 11 to 12, but it meant I had to take a dependency on some Duende pre-release versions that also use AutoMapper.   The Udap.Idp.Admin seems to work fine again.  Of course .NET 7 is pre-release but not after Nov 8th or 10th or something like that. :)