Preparing for a GCP release
Dropping OpenApi for now.  When enabling for net7.0 it was just a nightmare to try to get working.  After all it was only valuable for launching .well-known/udap.  The FHIR resources are not parsing to Open API (known gap).  Maybe later.
Adding RateLImiting before cloud push.
Adding a web_benchmark folder.  The restClient.http in combination with VS Code and the "REST Client" plugin makes for  nice dashboard simply test rest endpoints.  Think of it as curl or a super light Postman.
The Readme.md points to a quick and easy tool to place a load on your rest endpoint while testing at your desktop.  It is called SuperBenchmarker.  Very hand cli tool.