Decided to drop the multi targeted web sites
It was just taking too much of my time to figure out how to launch docker from Visual Studio.  If it is needed I can solve that later.

Adding community convenience links via new routes in the UDAP metadata controller.  See the tye.yaml for routes that will be on the dashboard.  This is just an easy way to find links to the UDAP community metadata endpoints with the specific community signed software statement