Working towards a GCP release.  Switching to SQL Server as the database.
Once inside a container on GCP "App Engine" feature the Sqlite technique could post a challenge.  Mostly I am new enough to cloud and GCP that I am not sure.  I bet that I could use shared storage but that is for another time.

Saving this issue for the next commit but here is a comment::
Switching to SQL Server, I ran into my first issue, in that ClientSecrets in the value column are larger than nvarchar(4000).  I am storing a x5c in there.  On Sqlite the column was defined at TEXT and I did not catch this issue.  I could spit each cert into its own entry rather than packaging in a x5c.  But I may some day run into this again.  So another option is to create a table called UdapClientSecrets.  Will have to look at this and see it I break any IdentityServer abstractions....