Misc work for passing udap.org test 18
Renamed UdapDB to UdapDB.SqlServer.  I was setting up Sqlite and SqlServer migrations and found that the last one to run was always the snapshot file winner when in the same assembly.  So I removed Sqlite.  Wasn't using it anyway.  But if I create Postgresql in the future I would want a new project for it.

Then setting RequirePkce to false when DCR via Authorization Code.  Remember we send in a signed software statement that is validated against the trust chain.  But it still seems PKCE might be a threat to revisit