Refactor:: Udap.Model project created
While adding a example Blazor Web Assembly (wasm) project and referencing the UdapMetadata it turned out that the project hosting it, is was referencing a runtime pack;  Microsoft.AspNetCore.App.  I chose to create a new project called Udap.Model that can host classes that act as models and constants.  Naming is hard and now I have projects with names I am not sure I like.  Saving a future refactor for project names like Udap.Client, Udap.Common and Udap.Model that don't exactly tell the story I like.  But for now this decoupling is necessary.