Started to build in the Certification and Endorsement data structure.
There is a lot of work ahead to exercise certification and endorsement.  But this is the base structure of building it.  More to come.  Just want to retain the work.  Added it to the the Udap.Org client test to see it in action.  Need to work on some extension methods to build the compact JWS structure.  There is a lot of laborious code examples to build repeated tests.  Again more work to come.