When resolving anchors, get all enabled unless the community name is passed.
The reality is there is no specification for passing community during registration.  So you have to get all of them to validated a registrations certificate to one of the community x509 chains.  At that point you would know which community the registration is for.  I have refactoring to do in the space.  So saving the identification of community membership for later.