Identity Server Pipeline testing feature integrated to better test UDAP
I delayed adding this in but it is a very very nice setup to make it much easier to test the upcoming /authorize? error response I need to fixup to behave different from the current redirect to /home/error?errorId=... pattern.  It is not MOQ but still pretty cool that there was a testing pattern I could leverage.

I refactored a few things.  I will probably migrate to a similar layout to the standard Duende project.  Like folder structure, where interfaces are vs implementations and project structure.  It just makes it easier to cross walk between projects.  After all udap-dotnet is just a few specialized components.  The spirit of this is to be a plugin to Duende.