Fixup test
I always forget that each of these test classes seeds a new Sqlite database.  I need to delete them so the get re-seeded.  On my Windows box I don't always do this because I forget.  Good thing is when it builds and tests run on the Linux GitHub agent it catches this.  I could delete the files every time a test runs but many times I am looking at the file with a tool and it is locked.