Simple component that helps with keeping session alive while a user is on the page but when not on the page it will let the session begin to expire.
It was annoying when I would be away from a page for more than an hour or the GCP server would spin down my instance and I would lose my session.  Yes just in process session.  But then I would click a button that generated  signed software statement only to find out my certificate was no long loaded.  This was on solution.  There are other techniques to guard this, but let's see how this works out.