Enable gcsfuse
This enables the second process called gcsfuse to run.  gcsfuse is a user-space file system for interacting with Google Cloud Storage.  So basically a mount point is mapped to a Bucket in Google Cloud.  This saves a lot o money.  The alternative FileStore solution running with the least resources started at a flat $204/month.  Remember that Fhirlabs is just a basic FHIR server using the files system as the database.