Bug: Fix:  Expect /authorize response errors to be transformed to redirects for missing client_id
This was coded but it expected a udap scope.  That makes no sense.  It was probably related to where my head was when I first created this.  I did not want to interrupt non udap traffic.  That is really not possible during this flow without a client_id.  I would need to have a client_id to lookup the client secrets to determine if they are a udap client.