Add Signed Algorithm choice and Subject Alt Name choice during registration in UdapEd UI Tool
Adding css classes to define the left and right columns in UdapEd.  Left is input and right is either a analyzed or knowledge styled paper.
Adjusted UdapDcrBuilderForClientCredentials and UdapDcrBuilderForAuthorizationCode so they could be inherited from.  Two new classes, suffixed with Unchecked were created so the UdapEd tool could get access to the underlying Document and set bad data so the tool can be used to experiment with incorrect registration requests.