wip UdapClient
Previous comment about creating a certificate resolver is not needed, at least not for that purpose of chasing AIA extensions.  This is actually handled by the .NET base libraries already when building a chain.

UdapClient is taking form.  The UdapEd UI tool has been updated to expose the new features of hooking events from the TrustChainValidator and JWT Validation.  More to come.