First cut of a Firely UDAP Plugin.
The Udap.Metadata.Vonk.Server plugin needs some packaging work.  It looks like it may be a npm package format?  Many adjustments to package references to survive a start up as a plugin.  :(  Also I need to take a stab at updating to HL7.Fhir.* assemblies to version 5.0  There is a big missmatch there once the repo layout and packaging was updated.  In the meantime I copied the old assemblies into my plugin folder.  Also the BouncyCastle.Crypto.dll and other Udap.*.dll files will need to be copied into the folder that is the plugins home.  Will come back to this later.