Require logo_uri during authorization code flow.  And registration key is now SAN and Community.  Grant type is not longer part of the client key.
Addd a LogRequired ServerSettings property.  Defaults to true and used in registration validation at auth server.
Exposed ClientRegistrationStore in UdapIdentityServerPipelin.  Not using it but left it exposed for future access in tests.
Started to break down the UdapDynamicClientRegistrationValidator so I could test outside integration.
Created a new Validators folder for validation tests.   Will revisit this and break down tests and validations later.