Registration replay attack prevention
I am not sure this is needed.  After all, all requests for an access token require the clients private asymmetric key.  I have seen another implementation that does require a unique jti for repeated registrations.  I have also seen implementations that do not require the jti.  The jti is optional in RFC 7519.

4.1.7.  "jti" (JWT ID) Claim

   The "jti" (JWT ID) claim provides a unique identifier for the JWT.
   The identifier value MUST be assigned in a manner that ensures that
   there is a negligible probability that the same value will be
   accidentally assigned to a different data object; if the application
   uses multiple issuers, collisions MUST be prevented among values
   produced by different issuers as well.  The "jti" claim can be used
   to prevent the JWT from being replayed.  The "jti" value is a case-
   sensitive string.  Use of this claim is OPTIONAL.

This by setting the ServerSetting.RegistrationJtiRequired to false the registration process will not require a jti and will not check replay attacks.

The build will break for the moment.  I need to upgrade Duende Identity Server packages to pick up a new IClock implmentation in the AddPluggableServices extension method.