jwks keys include UDAP issued certificates as optional signing credentials and validation keys
The goal is to issued Identity Tokens signed with a certificate that can be validated against a trusted anchor.  The ultimate goal after that is to add a certificate policy engine to assert identity assurance levels are issued by an authority we trust and can validate there certification to assert assurance levels.