Run dotnet tye using docker images
Running FhirLabsApi (FHIR Server), Udap.Auth.Server, Udap.Identity.Provider, and Udap.Identity.Provider.1 as Docker images.  Left UdapEd and Udap.Certificate.Server running as local services.

GCP deploy Docker Images were moved to Dockerfile.gcp for now.  I still would like to spend more time on this maybe removing the .gcp files, but busy on other items at the moment so accepting this technical debt for another day.

The push for this is to demonstrate how we could spin up an enterprise in Docker locally so other using this RI could see it in action sooner than later.

On a new system run the Udap.PKI.Generator tests.  First run the MakeCaWithIntermediateUdapAndSSLForDefaultCommunity to generate the certs needed for SSL.  On Windows run as admin so the CA and SubCA can be installed in Windows cert store.  If you can the comment out the calls to UpdateWindowsMachineStore and manually put them into your user store.  You shouldn't need to run this again for two years when the Certs expire.

Second run MakeCaWithIntermediateUdapForLocalhostCommunity to generate the UDAP test certs for local development.  If you do this and run the UdapServer.Tests, note you will have to delete the temp SQLite databases used for testing.  I just delete the bin folder in UdapServer.Tests.