Scope expander:: Deadend
This is not working out.  ParseScopeValues in UdapSmarV2SclopeParser ends up with sclopes in the Request?.ValidatedResources?.RawScopeValues.  So in CreateCodeFlowResponseAsync of UdapAuthorizeResponseGenerator we are faced with a readon only property in AuthorizeResponse.Scope.  The thing is, the read only property is a good thing.  We shouldn't have a way to accidently or maliciously change scopes.  I am going to commit this code even though it doesn't compile.  Going to take a different route on my next try.