SmartV2 Expansion : Round 2
This time I isolated the expansion to the installation of database and during registration.  This is much simpler.  Although the database could have a large number of ApiScopes.  Sticking with this technique for now.

Still need to implement wild card scope expansion as part of the use case we demonstrated at the 2023 CMS connectathon.  More to come.