UDAP Dynamic Provider tests are mechanically functional
Need to commit code to plant a stake.    It is not 100% done.  I still  need a OidcProvider registered as a template to hook TieredOAuthenticationHandler.  I have not persisted it yet.  Need to extend the integration test to persist and then re-run process asserting the provider is persisted.  Persistence in Duende IdentityServer means persisting the provider data to the Provider table.

One noteworthy friction point:  I have to set the Scope property of the templated OidcProvider.  Even though the TieredOAuthAuthenticationOptions is registered and preset to scopes openid, email and profile the Duende mechanism for dynamically loading profiles changes the Scope property via post configuration process just before I go to use it.  It seems like a bug, but would take some time to build a simple PR or why.  But in the mean time setting the Scope in the OidcProvider template instance resolves the concern.

All in all there is a lot to keep in your head to develop this dynamic provider code.  But the consumer experience when building out a UDAP enabled IdentityServer should prove to be easier to setup than the current implementation.