Dynamic Tiered OAuth.  Unwind the Duende dynamic provider technique.
After some time experimenting with Duende dynamic provider, I realized I did not need it for my dynamic behavior for registering with unknown UDAP enabled identity services.  I just need a single entry like this:
builder.Services.AddAuthentication()
            .AddTieredOAuth(options =>
            {
                options.SignInScheme = IdentityServerConstants.ExternalCookieAuthenticationScheme;
            });
When the client app supplies a idp query param the TieredOAuthHandler is triggered.  It auto registers and then creates an TieredClient entry in the DB.  This reduced  lot of code.  It was a long journey to get here.
May start to transform TieredOAuthHandler from a OAuthHandler implementation to a OpendIdHandler.  The OpenIdHandler has more hooks for thing like federated signout.