Building UdapClient Registration interfaces.
One usage for Auth Code flow in TieredAuthTests when registering the client.

Next will be client credentials flow.

The client is getting close to complete.  Will not use all these features in UdapEd because of the way UdapEd is meant to break down the process for diagnostics and visualization of the UDAP process.