Fixup Authorize error flows
Back in line with RFC 6749 and udap.org UDAPTestTool.
Will have to deploy and test to validate, but I think this is going in the right direction.  Duende by default redirects to the server error page.  This behavior isn't in pattern with UDAPTestTool and that is why the UdapAuthorizationResponseMiddleware component exists.