Validation enhancments
Updating validation to check alg and x5c failures.  Found that I had a fundamental error in that I was not configuring the TokeValidationParameters.VaidAlgorithms from a known list.  Rather I was setting it from the incoming alg which could be null.  This was bad.

Also some changes in udap.org/TestTool could lead to removing the removing the ServerSupport setting in SeverSettings.  Not sure if I can completely remove it yet.  Seems it might have to stay because the TestTool registers with no scopes if I remember right.