New tests and fixups to handle Authorization Certificate Rolloer
A client registers and two entries are set in the Secrets table.  Those entries contain an expires column that is the same as the UDAP client certificate expiration.  Some time in the future the client Authorizes with a signed client_assertion.  On the IdentityServer side the secrets are resolved from the secret table.  But because they are expired the secrets are not available to the UdapJwtSecretValidator.   The validator needs to handle this situation by informing the Client Store to updated the secrets in supplied by the signed client_assertion.  Of course this should not happen until the signed client_assertion has been validated with the JsonWebTokenHandler.  Once this is done the expiration dates for the existing secrets can be updated and then the lookups for anchors related to the community can be resolved and finally the x509 Chaining process can be validated.