UserPerson converted to a JsonElement
Another method could have been to change it to an HL7 Resource or Person type.  But then I would need to serialize that property using the FhirJsonSerializer.  But I don't really have access to the serialization.  Thee serialization happens deep in HttpMessageInvoker.RequestTokenAsync(ProtocolRequest request) .  Maybe there is a way, but not ready to attack that.  And if I did then I might have to re-implement HttpClientTokenRequestExtensions.  Not sure.  And maybe this will work.  Until I consume this as a server I am not sure if there is still more work here.