Simplify Authorization Extension Object Serialization structures
No longer inheriting from Dictionary<string, object>.  This also means it is not easy to extend a Authorization Extension Object with extra claims.  You can still of course add a claim to the root of the JwtPayload.  After all if you added a claim to an existing defined Auth Ext Object it wouldn't carry any meaning when received at the authorization server.  I am hoping this plays out well when consumed in UdapEd, because up to this point it has been very difficult to consume the previous Auth Ext Objects.