Removing the TefcaMetdata
I previously mis-interpreted the TEFCA SOP for facilitated FHIR.  I thought there were new metadata items to server from Discovery but it is actually new client metadata.  The language is not explicit.  I did not unwind all the work to make Metadata extensible.  As this could be a requirement in the future.  But for now I wanted to clean this up.