Working on the ability to terminate trust at a chosen Trust Anchor.
This should pass the tests.
It used to be that we did not fail when chainBuilder.Build(certificate) returned false.  That is not longer the case and I think it is a better way to validate.  In the past we would examine the problem flags to determine if it was a failure.

This means many tests need to run with X509RevocationMode set to NoCheck and that is fine for these tests.