I don't know why the builds are failing
I have tried to remove the anchor locally on my Windows box from everywhere it is cached.  On windows I can see in the logs that is resolves the CA that is on the internet based on Thumbprint of 80E318ABFB0FC53092FDAA593D024D7173EC40C8.  The thing is that is not the CA that chains up to the end cert after dynamically generating during builds.  I have tested this locally on Ubuntu 22.04.  I do see that on GitHub the builds are running on 24.04.  So I am downloading 24.04 to run in WSL to see if I can get the same failure here.  This commit is nothing more than being curious what will happen.